+++ packtools/stylechecker.py
@@ -23,6 +23,13 @@ import packtools
 logger = logging.getLogger(__name__)
 
 
+EPILOG = """\
+Copyright 2013 SciELO <scielo-dev@googlegroups.com>.
+Licensed under the terms of the BSD license. Please see LICENSE in the source
+code for more information.
+"""
+
+
 class XMLError(Exception):
     """ Represents errors that would block XMLValidator instance from
     being created.
@@ -54,15 +61,18 @@ def prettify(jsonobj, colorize=True):
 def get_xmlvalidator(xmlpath, no_network, extra_sch):
     try:
         parsed_xml = packtools.XML(xmlpath, no_network=no_network)
-    except IOError as e:
+    except IOError as exc:
+        logger.debug(exc)
         raise XMLError('Error reading %s. Make sure it is a valid file-path or URL.' % xmlpath)
-    except etree.XMLSyntaxError as e:
-        raise XMLError('Error reading %s. Syntax error: %s' % (xmlpath, e))
+    except etree.XMLSyntaxError as exc:
+        logger.debug(exc)
+        raise XMLError('Error reading %s. Syntax error: %s' % (xmlpath, exc))
 
     try:
         xml = packtools.XMLValidator(parsed_xml, extra_schematron=extra_sch)
-    except ValueError as e:
-        raise XMLError('Error reading %s. %s.' % (xmlpath, e))
+    except ValueError as exc:
+        logger.debug(exc)
+        raise XMLError('Error reading %s. %s.' % (xmlpath, exc))
 
     return xml
 
@@ -106,18 +116,20 @@ def summarize(validator, assets_basedir=None):
 
 
 @packtools.utils.config_xml_catalog
-def main():
+def _main():
 
     packtools_version = pkg_resources.get_distribution('packtools').version
 
-    parser = argparse.ArgumentParser(description='stylechecker cli utility')
+    parser = argparse.ArgumentParser(
+            description='SciELO PS stylechecker command line utility.',
+            epilog=EPILOG)
     parser.add_argument('--annotated', action='store_true',
                         help='reproduces the XML with notes at elements that have errors')
     parser.add_argument('--nonetwork', action='store_true',
                         help='prevents the retrieval of the DTD through the network')
     parser.add_argument('--assetsdir', default=None,
                         help='lookup, at the given directory, for each asset referenced by the XML. current working directory will be used by default.')
-    parser.add_argument('XML', nargs='+',
+    parser.add_argument('XML', nargs='*',
                         help='filesystem path or URL to the XML')
     parser.add_argument('--version', action='version', version=packtools_version)
     parser.add_argument('--loglevel', default='WARNING')
@@ -127,19 +139,21 @@ def main():
                         help='runs an extra validation using an external schematron schema.')
     args = parser.parse_args()
 
-    logging.basicConfig(level=getattr(logging, args.loglevel))
+    logging.basicConfig(level=getattr(logging, args.loglevel.upper()))
 
     print('Please wait, this may take a while...', file=sys.stderr)
 
+    input_args = args.XML or sys.stdin
     summary_list = []
-    for xml in packtools.utils.flatten(args.XML):
+
+    for xml in packtools.utils.flatten(input_args):
         logger.info('starting validation of %s' % (xml,))
 
         try:
             xml_validator = get_xmlvalidator(xml, args.nonetwork, args.extrasch)
             logger.debug('XMLValidator repr: %s' % repr(xml_validator))
-        except XMLError as e:
-            logger.debug(e)
+        except XMLError as exc:
+            logger.debug(exc)
             logger.warning('Error validating %s. Skipping. Run with DEBUG for more info.', xml)
             continue
 
@@ -164,13 +178,17 @@ def main():
                     assets_basedir = args.assetsdir or os.path.dirname(xml)
 
                 summary = summarize(xml_validator, assets_basedir=assets_basedir)
-            except TypeError as e:
-                logger.debug(e)
-                logger.warning('Error validating %s. Skipping. Run with DEBUG for more info.', xml)
+            except TypeError as exc:
+                logger.debug(exc)
+                logger.warning(
+                        'Error validating %s. Skipping. Run with DEBUG for more info.',
+                        xml)
                 continue
 
             summary['_xml'] = xml
-            summary['is_valid'] = bool(xml_validator.validate()[0] and xml_validator.validate_style()[0])
+            summary['is_valid'] = bool(
+                    xml_validator.validate()[0] and
+                    xml_validator.validate_style()[0])
 
             summary_list.append(summary)
 
@@ -179,6 +197,13 @@ def main():
     if summary_list:
         print(prettify(summary_list, colorize=args.nocolors))
 
+
+def main():
+    try:
+        _main()
+    except KeyboardInterrupt:
+        logger.debug('The program is terminating due to SIGTERM.')
+        pass
 
 if __name__ == '__main__':
     main()